Introduce RestResourceErrorMapper for bean configuration
